move zeitwerk validation support into main package
